Move CLI parsing functions to their own module
